Migrate non-apply part of shared to TypeScript
