--- index.js
@@ -1,3 +1,0 @@
-require('dotenv').load();
-
-require('js/server');
