+++ src/index.ts
@@ -1,0 +1,5 @@
+import * as dotenv from 'dotenv';
+
+dotenv.load();
+
+require('js/server');
