+++ src/js/client/live.ts
@@ -2,7 +2,7 @@ import * as $ from 'jquery';
 import * as Pusher from 'pusher-js';
 import * as moment from 'moment';
 
-import * as Countdown from 'js/shared/countdown';
+import { Countdown } from 'js/shared/countdown';
 
 // Add globals for Pusher config
 declare global {
