--- src/js/shared/colors.js
@@ -1,1 +1,0 @@
-exports.PRIMARY_COLOR = '#f93822';
