+++ src/js/shared/colors.ts
@@ -1,0 +1,1 @@
+export const PRIMARY_COLOR = '#f93822';
